+++ src/mockfn/core.clj
@@ -2,11 +2,11 @@
   (:require [mockfn.macros :as macros]
             [mockfn.matchers :as matchers]))
 
-(def #^{:macro true} providing #'macros/providing)
-(def #^{:macro true} verifying #'macros/verifying)
+(def #^{:macro true} ^:deprecated providing #'macros/providing)
+(def #^{:macro true} ^:deprecated verifying #'macros/verifying)
 
-(def exactly matchers/exactly)
-(def at-least matchers/at-least)
-(def at-most matchers/at-most)
-(def any matchers/any)
-(def a matchers/a)
+(def ^:deprecated exactly matchers/exactly)
+(def ^:deprecated at-least matchers/at-least)
+(def ^:deprecated at-most matchers/at-most)
+(def ^:deprecated any matchers/any)
+(def ^:deprecated a matchers/a)