+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project are documented in this file.
 
 ## Releases
 
+### 0.2.0 - Unreleased
+- Added syntax sugar for clojure.test.
+- Deprecated everything in the `mockfn.core` namespace.
+
 ### 0.1.0 - 2018-03-18
 - Added function stubs.
 - Added call verification.
@@ -12,4 +16,3 @@ All notable changes to this project are documented in this file.
 ## Planned
 
 - Chained responses
-- Syntax sugar deftest